folsetech-ai-studio/ai/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
openai-whisper>=20230314
demucs>=4.0.0
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
